feat(design): remove error message dependency on form-field
