+++ libs/design/src/atoms/form/error-message/error-message.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../scss/daff-util';
+@import '../../../scss/daff-util';
 
 .daff-error-message {
 	display: block;
