+++ libs/design/src/atoms/form/error-message/error-message.component.ts
@@ -1,12 +1,12 @@
-import { Component, ViewEncapsulation } from '@angular/core';
+import { Component, ViewEncapsulation, HostBinding } from '@angular/core';
 
 @Component({
   selector: 'daff-error-message',
   template: '<ng-content></ng-content>',
   styleUrls: ['./error-message.component.scss'],
-  host: {
-    'class': 'daff-error-message'
-  },
   encapsulation: ViewEncapsulation.None
 })
-export class DaffErrorMessageComponent {}
+export class DaffErrorMessageComponent {
+  
+  @HostBinding('class.daff-error-message') class = true;
+}
