+++ libs/design/src/atoms/form/error-message/error-message.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { DaffErrorMessageComponent } from './error-message.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  exports: [
+    DaffErrorMessageComponent
+  ],
+  declarations: [
+    DaffErrorMessageComponent
+  ]
+})
+export class DaffErrorMessageModule { }
