+++ libs/design/src/atoms/form/form-field/form-field/form-field.component.spec.ts
@@ -4,7 +4,7 @@ import { By } from '@angular/platform-browser';
 import { ReactiveFormsModule, FormControl, Validators } from '@angular/forms';
 
 import { DaffFormFieldComponent } from './form-field.component';
-import { DaffErrorMessageComponent } from '../error-message/error-message.component';
+import { DaffErrorMessageComponent } from '../../error-message/error-message.component';
 import { DaffFormFieldMissingControlMessage } from '../form-field-errors';
 import { DaffFormFieldControl } from '../form-field-control';
 import { DaffInputModule } from '../../input/public_api';
