+++ libs/design/src/atoms/form/form-field/public_api.ts
@@ -1,4 +1,3 @@
 export { DaffFormFieldModule } from './form-field.module';
 export * from './form-field/form-field.component';
 export { DaffFormFieldControl } from './form-field-control';
-export * from './error-message/error-message.component';
