+++ libs/design/src/scss/theming/_theme.scss
@@ -1,5 +1,5 @@
 @import '../../atoms/button/button-theme';
-@import '../../atoms/form/form-field/error-message/error-message-theme';
+@import '../../atoms/form/error-message/error-message-theme';
 @import '../../atoms/form/form-field/form-field/form-field-theme';
 @import '../../atoms/form/input/input/input-theme';
 @import '../../atoms/form/select/select/select-theme';
